Reorganized code, quicker linear interpolation
